pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toxicity-agent"
version = "0.1.0"
description = "End-to-end Hate Speech & Toxicity Detection system with agentic moderation and deployment"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Your Name"}]

dependencies = [
  "torch>=2.1.0",
  "transformers>=4.40.0",
  "datasets>=2.19.0",
  "accelerate>=0.30.0",
  "scikit-learn>=1.3.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "tqdm>=4.66.0",
  "joblib>=1.3.0",
  "matplotlib>=3.7.0",
  "detoxify==0.5.2",
  "sentencepiece>=0.1.99",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.6.0",
  "PyYAML>=6.0.1",
  "python-dotenv>=1.0.1",
  "langdetect>=1.0.9",
  "reportlab>=4.0.0",
  "python-pptx>=0.6.23",
  "Pillow>=9.5.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "httpx>=0.26.0"]

[project.scripts]
toxicity-agent = "toxicity_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
